fix: Correção de bug nos valores indisponíveis
Correção de bug nos valores indisponíveis, onde o número estava sendo convertido para INT, onde deveria ser convertido para FLOAT.

+++ app/Http/Controllers/ContrachequeController.php
@@ -192,7 +192,7 @@ class ContrachequeController extends Controller
                 $tipos = $value['tipo'] == 1 ? 'receitas' : 'descontos';
                 $abrev = strtolower($value['descricao']);
                 $abrev = '_' . str_replace(" ", "_", $abrev);
-                $valor = intval($value['valor']);
+                $valor = floatval($value['valor']);
                 $rubrica = strtoupper($value['descricao']);
                 $ir = $value['tributavel'] == 1 ? 'ir' : 'n_ir';
                 $adicNatal = '13N';
